backend/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Menna"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.2"}
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
